fixed being able to upload non images

+++ routes/user-routes.js
@@ -42,8 +42,13 @@ router.post('/upload', authCheck, urlencodedParser, upload.single('photo'), (req
     var imagename = uniqid() + '.png';
     var path = "public/uploads/";
     var storage_path = "/uploads/";
+    var isjpeg = true;
     if (req.file){
-        console.log("file = to this = " + req.file.path);
+        console.log("file contains = " + req.file.mimetype);
+        if (req.file.mimetype != 'image/jpeg' && req.file.mimetype != 'image/png') {
+            isjpeg = false;
+            console.log('NOT AN JPEG IMAGE OR PNG IMAGE');
+        }
         var imagedest1 = req.file.path;
         //res.json(req.file);
     } else {
@@ -69,53 +74,55 @@ router.post('/upload', authCheck, urlencodedParser, upload.single('photo'), (req
     var images = [imagedest1, 'public/assets/'+req.body.sticker0, 'public/assets/'+req.body.sticker1, 'public/assets/'+req.body.sticker2];
 
     var jimp_imgs = [];
+    if (isjpeg === true){
 
-    for (var i = 0; i < images.length; i++){
-        if (images[i] != 'public/assets/none') {
-            jimp_imgs.push(jimp.read(images[i]));
-            console.log('image number '+i+' from images: '+images[i]);
+        for (var i = 0; i < images.length; i++){
+            if (images[i] != 'public/assets/none') {
+                jimp_imgs.push(jimp.read(images[i]));
+                console.log('image number '+i+' from images: '+images[i]);
+            }
         }
-    }
-
-    Promise.all(jimp_imgs).then((data) =>{
-        return Promise.all(jimp_imgs);
-    }).then((data) => {
-        let i = 1;
-        //console.log(data[0]);
-
-        let width = data[0].bitmap.width;
-        let height = data[0].bitmap.height;
-        //data[0].resize(1280, 898);
-        while (data[i]) {
-            data[i].resize(width, height);
-            //console.log(data[i]);
-            data[0].composite(data[i],0,0);
-            i++;
-        }
-        data[0].write(path+imagename, (err) =>{
-            if (err)
-                throw err
-            else
-            console.log('jimp image created');
-        })
-        fs.unlink(imagedest1, (err) => {
-            if (!err)
-                console.log('file deleted');
-            else
-                throw err;
-        });
-
-        
-
-        // add to database
-        new Post({
-            userid: req.user._id,
-            image: storage_path+imagename
-        }).save().then((newpost) => {
-            console.log('new post created with info: ' + newpost)
-        });
-    });
     
+        Promise.all(jimp_imgs).then((data) =>{
+            return Promise.all(jimp_imgs);
+        }).then((data) => {
+            let i = 1;
+            //console.log(data[0]);
+    
+            let width = data[0].bitmap.width;
+            let height = data[0].bitmap.height;
+            //data[0].resize(1280, 898);
+            while (data[i]) {
+                data[i].resize(width, height);
+                //console.log(data[i]);
+                data[0].composite(data[i],0,0);
+                i++;
+            }
+            data[0].write(path+imagename, (err) =>{
+                if (err)
+                    throw err
+                else
+                console.log('jimp image created');
+            })
+            fs.unlink(imagedest1, (err) => {
+                if (!err)
+                    console.log('file deleted');
+                else
+                    throw err;
+            });
+    
+            
+    
+            // add to database
+            new Post({
+                userid: req.user._id,
+                image: storage_path+imagename
+            }).save().then((newpost) => {
+                console.log('new post created with info: ' + newpost)
+            });
+        });
+        
+    }
     //console.log('req user has: '+ req.user);
     //User.findById(req.user)
     res.render('post', {user: req.user});    
@@ -135,7 +142,6 @@ router.post('/commentnlike/:post/:op', authCheck, urlencodedParser, (req, res) =
                     userid: req.user._id,
                     comment: req.body.comment_text
                 });
-                console.log('comment : +' + req.body.comment_text);
                 current_post.save().then((data) =>{
                     var names = [];
                     for(i = 0; current_post.comments[i]; i++){
@@ -149,7 +155,6 @@ router.post('/commentnlike/:post/:op', authCheck, urlencodedParser, (req, res) =
                     }
 
                     User.findOne({_id: current_post.userid}).then((post_creator) =>{
-                        console.log('done names: '+ names);
                         if (post_creator){
                             res.render('comments', {user: req.user, post: current_post, names: names, post_creator: post_creator.username});
                         } else {
@@ -171,6 +176,8 @@ router.get('/commentnlike/:post/:op', authCheck, (req, res) =>{
         Post.findOne({_id: req.params.post}).then((current_post) =>{
             if (current_post){
                 var names = [];
+                
+                // This part ZB is just so that i can find the name of the user since i keep my user and post models separate
                 for(i = 0; current_post.comments[i]; i++){
                     User.findOne({_id: current_post.comments[i].userid}).then((result) => {
                         var comment_username = result.username;
@@ -179,7 +186,7 @@ router.get('/commentnlike/:post/:op', authCheck, (req, res) =>{
                         );
                     });
                 }
-
+                
                 User.findOne({_id: current_post.userid}).then((post_creator) =>{
                     if (post_creator){
                         res.render('comments', {user: req.user, post: current_post, names: names, post_creator: post_creator.username});
@@ -196,7 +203,6 @@ router.get('/commentnlike/:post/:op', authCheck, (req, res) =>{
                 var existing = false;
                 for (i = 0; current_post.likes[i]; i++){
                     if (current_post.likes[i].userid == userid){
-                        console.log('found similar like');
                         existing = true;
                         break ;
                     }
@@ -206,15 +212,10 @@ router.get('/commentnlike/:post/:op', authCheck, (req, res) =>{
                     current_post.likes.push({
                         userid: userid
                     });
-                    console.log('add like');
                     current_post.save();
                 } else {
                     // note to self this is working fine, it seems as if the page was holding on to what ever information it had in the past (cache?)
-                    Post.updateOne({_id: req.params.post}, {$pull : {likes: {userid: userid}}}, {multi: true}).then((result) => {
-                        if (result){
-                            console.log('rem like');
-                        }
-                    }); 
+                    Post.updateOne({_id: req.params.post}, {$pull : {likes: {userid: userid}}}, {multi: true})
                 } 
 
                 // this should add the names according to their id's and push them into the array
@@ -254,8 +255,6 @@ router.get('/viewposts/:page', authCheck, (req, res) =>{
             } else {
                 var page = 0;
             }
-            console.log('current user post :'+current_user_post+')))end(((')
-            console.log('page = ' + page);
             res.render('viewposts', {user: req.user, posts: current_user_post, page: page});
         } else {
             console.log('something went wrong with finding the users posts');
@@ -264,12 +263,15 @@ router.get('/viewposts/:page', authCheck, (req, res) =>{
     });
 });
 
-router.get('/viewposts/delete/:id', authCheck, (req, res) =>{
+router.get('/viewposts/delete/:id/:page', authCheck, (req, res) =>{
     //res.send('you are logged in user ' + req.user.username);
     Post.findOneAndDelete({userid: req.user._id, _id: req.params.id}).then((current_user_post) =>{
         if (current_user_post) {
-            console.log('current user post :'+current_user_post+')))end(((')
-            console.log('page = ' + page);
+            if (req.params.page) {
+                var page = parseInt(req.params.page, 10);
+            } else {
+                var page = 0;
+            }
             res.render('viewposts', {user: req.user, posts: current_user_post, page: page});
         } else {
             res.redirect('0');
